refactor: extract process ledger to LedgerManager
