+++ packages/server-wallet/src/wallet/wallet.ts
@@ -34,7 +34,6 @@ import {ValidationErrorItem} from 'joi';
 
 import {Bytes32} from '../type-aliases';
 import {createLogger} from '../logger';
-import * as ProcessLedgerQueue from '../protocols/process-ledger-queue';
 import * as UpdateChannel from '../handlers/update-channel';
 import * as JoinChannel from '../handlers/join-channel';
 import * as ChannelState from '../protocols/state';
@@ -60,8 +59,8 @@ import {
 import {DBAdmin} from '../db-admin/db-admin';
 import {WALLET_VERSION} from '../version';
 import {ObjectiveManager} from '../objectives';
-import {Channel} from '../models/channel';
 import {SingleAppUpdater} from '../handlers/single-app-updater';
+import {LedgerManager} from '../protocols/ledger-manager';
 
 import {Store, AppHandler, MissingAppHandler} from './store';
 import {
@@ -95,6 +94,7 @@ export class SingleThreadedWallet
   store: Store;
   chainService: ChainServiceInterface;
   objectiveManager: ObjectiveManager;
+  ledgerManager: LedgerManager;
   logger: Logger;
 
   readonly walletConfig: ServerWalletConfig;
@@ -147,6 +147,12 @@ export class SingleThreadedWallet
     this.objectiveManager = ObjectiveManager.create({
       store: this.store,
       chainService: this.chainService,
+      logger: this.logger,
+      timingMetrics: this.walletConfig.metricsConfiguration.timingMetrics,
+    });
+
+    this.ledgerManager = LedgerManager.create({
+      store: this.store,
       logger: this.logger,
       timingMetrics: this.walletConfig.metricsConfiguration.timingMetrics,
     });
@@ -663,112 +669,11 @@ export class SingleThreadedWallet
 
     const ledgersToProcess = _.uniq(ledgerIdsFromChannels.concat(ledgerIdsFundingChannels));
 
-    while (ledgersToProcess.length) {
-      const ledgerChannelId = ledgersToProcess[0];
+    ledgersToProcess.forEach(async ledgerChannelId => {
+      const result = await this.ledgerManager.crank(ledgerChannelId, response);
+      requiresAnotherCrankUponCompletion = requiresAnotherCrankUponCompletion || result;
+    });
 
-      await this.store.lockApp(ledgerChannelId, async (tx, channel) => {
-        const setError = (e: Error) => tx.rollback(e);
-
-        const markLedgerAsProcessed = () => ledgersToProcess.shift();
-
-        // TODO: Move these checks inside the DB query when fetching ledgers to process
-        if (!channel.protocolState.supported || channel.protocolState.supported.turnNum < 3) {
-          markLedgerAsProcessed();
-          return;
-        }
-
-        const pessimisticallyAddStateAndProposalToOutbox = () => {
-          const {
-            fundingChannel: {myIndex, channelId, participants, latestSignedByMe, supported},
-            myLedgerProposal: {proposal, nonce},
-          } = protocolState;
-          if (latestSignedByMe && supported) {
-            /**
-             * Always re-send a proposal if I have one withstanding, just in case.
-             */
-            if (proposal)
-              response.queueProposeLedgerUpdate(channelId, myIndex, participants, proposal, nonce);
-            /**
-             * Re-send my latest signed ledger state if it is not supported yet.
-             */
-            if (latestSignedByMe.turnNum > supported.turnNum)
-              response.queueState(latestSignedByMe, myIndex, channelId);
-          }
-        };
-
-        const protocolState = await ProcessLedgerQueue.getProcessLedgerQueueProtocolState(
-          this.store,
-          ledgerChannelId,
-          tx
-        );
-        const action = recordFunctionMetrics(
-          ProcessLedgerQueue.protocol(protocolState),
-          this.walletConfig.metricsConfiguration.timingMetrics
-        );
-
-        if (!action) {
-          markLedgerAsProcessed();
-          if (!requiresAnotherCrankUponCompletion) pessimisticallyAddStateAndProposalToOutbox();
-          response.queueChannelState(protocolState.fundingChannel);
-        } else {
-          try {
-            switch (action.type) {
-              case 'DismissLedgerProposals': {
-                await this.store.removeLedgerProposals(ledgerChannelId, tx);
-                requiresAnotherCrankUponCompletion = true;
-                return;
-              }
-
-              case 'SignLedgerUpdate': {
-                const {myIndex, channelId} = protocolState.fundingChannel;
-                const channel = await Channel.forId(channelId, tx);
-                const signedState = await this.store.signState(channel, action.stateToSign, tx);
-                response.queueState(signedState, myIndex, channelId);
-                return;
-              }
-
-              case 'ProposeLedgerUpdate': {
-                // NOTE: Proposal added to response in pessimisticallyAddStateAndProposalToOutbox
-                await this.store.storeLedgerProposal(
-                  action.channelId,
-                  action.outcome,
-                  action.nonce,
-                  action.signingAddress,
-                  tx
-                );
-                return;
-              }
-
-              case 'MarkInsufficientFunds': {
-                await this.store.markLedgerRequests(action.channelsNotFunded, 'fund', 'failed', tx);
-                return;
-              }
-
-              case 'MarkLedgerFundingRequestsAsComplete': {
-                const {fundedChannels, defundedChannels, ledgerChannelId} = action;
-
-                /**
-                 * After we have completed some funding requests (i.e., a new ledger state
-                 * has been signed), we can confidently clear now-stale proposals from the DB.
-                 */
-                await this.store.removeLedgerProposals(ledgerChannelId, tx);
-
-                await this.store.markLedgerRequests(fundedChannels, 'fund', 'succeeded', tx);
-                await this.store.markLedgerRequests(defundedChannels, 'defund', 'succeeded', tx);
-
-                requiresAnotherCrankUponCompletion = true;
-                return;
-              }
-              default:
-                throw 'Unimplemented';
-            }
-          } catch (err) {
-            this.logger.error({err}, 'Error handling action');
-            await setError(err);
-          }
-        }
-      });
-    }
     return requiresAnotherCrankUponCompletion;
   }
